fixing up built in player
